Some much-needed minor improvements.
Change the default behaviour to Big endian; the Autodetect really only works part of the time, and it makes things unpredictable. Plus, it's very seldom that one finds a file that isn't Big endian.
Also, change the mmap mode to not require write access. This makes reading things much easier on shared systems.